requirements.txt
wheel==0.33.6
joblib
pptk
scipy==1.4.1
numpy==1.18.5
numba
pyyaml
tensorflow-gpu==2.3
tensorflow-addons==0.11.2
tensorflow-probability==0.11.1
flatten-dict
h5py==2.10.0
pytest
matplotlib
pypng
opencv-python
datadiff
psutil
pykitti
